+++ Editor/Model/MemoryReport.cs
@@ -132,6 +132,6 @@ namespace GameDistrict.MemoryShield.Model
 
     public static class MemoryShieldVersion
     {
-        public const string Version = "1.0.1";
+        public const string Version = "1.0.2";
     }
 }